+++ jgclark.Dashboard/src/react/components/ComboBox.jsx
@@ -1,3 +1,5 @@
+// dbw Note: Maybe not used anymore. Trying to move all to the DashboardDialog version
+
 // @flow
 //--------------------------------------------------------------------------
 // Dashboard React component to show an HTML ComboBox control, with various possible settings.
@@ -21,7 +23,7 @@ type ComboBoxProps = {
   onChange: (value: string) => void,
   inputRef?: { current: null | HTMLInputElement },
   compactDisplay?: boolean,
-};
+}
 
 //--------------------------------------------------------------------------
 // ComboBox Component Definition
@@ -42,13 +44,17 @@ const ComboBox = ({ label, options, value, onChange, inputRef, compactDisplay = 
   /**
    * Toggles the dropdown open or closed.
    */
-  const toggleDropdown = () => setIsOpen(!isOpen)
+  const toggleDropdown = () => {
+    logDebug(`Dashboard toggleDropdown, isOpen: ${String(isOpen)}`)
+    setIsOpen(!isOpen)
+  }
 
   /**
    * Handles the selection of an option.
    * @param {string} option - The selected option.
    */
   const handleOptionClick = (option: string) => {
+    logDebug(`Dashboard handleOptionClick, option: ${option}`)
     setSelectedValue(option)
     onChange(option)
     setIsOpen(false)
@@ -67,49 +73,58 @@ const ComboBox = ({ label, options, value, onChange, inputRef, compactDisplay = 
   /**
    * Scrolls the combobox fully into view.
    */
-  const handleComboboxOpen = () => {
-    setTimeout(() => {
-      if (comboboxInputRef.current instanceof HTMLInputElement) {
-        comboboxInputRef.current.scrollIntoView({ behavior: 'smooth', block: 'end' })
-        console.log('Found comboboxInputRef so added scroll handler')
-      } else {
-        console.log('Could not find comboboxInputRef')
-      }
-    }, 100) // Delay to account for rendering/animation
-  }
+  // const handleComboboxOpen = () => {
+  //   setTimeout(() => {
+  //     if (comboboxInputRef.current instanceof HTMLInputElement) {
+  //       comboboxInputRef.current.scrollIntoView({ behavior: 'smooth', block: 'end' })
+  //       console.log('Found comboboxInputRef so added scroll handler')
+  //     } else {
+  //       console.log('Could not find comboboxInputRef')
+  //     }
+  //   }, 100) // Delay to account for rendering/animation
+  // }
 
   //----------------------------------------------------------------------
   // Effects
   //----------------------------------------------------------------------
 
   useEffect(() => {
-    document.addEventListener('mousedown', handleClickOutside)
+    if (isOpen) {
+      logDebug(`Dashboard ComboBox useEffect: Adding mousedown listener`)
+      document.addEventListener('mousedown', handleClickOutside)
+    } else {
+      logDebug(`Dashboard ComboBox useEffect: Removing mousedown listener`)
+      document.removeEventListener('mousedown', handleClickOutside)
+    }
+
+    // Cleanup function to ensure the listener is removed when the component unmounts or isOpen changes
     return () => {
       document.removeEventListener('mousedown', handleClickOutside)
     }
-  }, [])
+  }, [isOpen])
 
-  useEffect(() => {
-    const combobox = comboboxInputRef.current
-    if (combobox instanceof HTMLInputElement) {
-      console.log('adding event listener for comboboxInputRef')
-      combobox.addEventListener('click', handleComboboxOpen)
-    }
-    return () => {
-      if (combobox instanceof HTMLInputElement) {
-        console.log('removing event listener for comboboxInputRef')
-        combobox.removeEventListener('click', handleComboboxOpen)
-      }
-    }
-  }, [])
+  // useEffect(() => {
+  //   const combobox = comboboxInputRef.current
+  //   if (combobox instanceof HTMLInputElement) {
+  //     console.log('adding event listener for comboboxInputRef')
+  //     combobox.addEventListener('click', handleComboboxOpen)
+  //   }
+  //   return () => {
+  //     if (combobox instanceof HTMLInputElement) {
+  //       console.log('removing event listener for comboboxInputRef')
+  //       combobox.removeEventListener('click', handleComboboxOpen)
+  //     }
+  //   }
+  // }, [])
 
-  useEffect(() => {
-    setSelectedValue(value)
-  }, [value])
+  // useEffect(() => {
+  //   setSelectedValue(value)
+  // }, [value])
 
   //----------------------------------------------------------------------
   // Render
   //----------------------------------------------------------------------
+  logDebug(`Dashboard ComboBox render, isOpen: ${String(isOpen)}`, { options })
   return (
     <div className={compactDisplay ? 'combobox-container-compact' : 'combobox-container'}>
       <label className="combobox-label">{label}</label>
@@ -125,11 +140,7 @@ const ComboBox = ({ label, options, value, onChange, inputRef, compactDisplay = 
         {isOpen && (
           <div className="combobox-dropdown" ref={comboboxRef}>
             {options.map((option: string) => (
-              <div
-                key={option}
-                className="combobox-option"
-                onClick={() => handleOptionClick(option)}
-              >
+              <div key={option} className="combobox-option" onClick={() => handleOptionClick(option)}>
                 {option}
               </div>
             ))}